beta 파일 수정, main js 분리 및 선택사항 일부 구현
