+++ beta/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="./css/index.css">
   <script src="./js/index.js"></script>
-  <script src="./js/arrow.js" defer></script>
+  <!-- <script src="./js/arrow.js" defer></script> -->
   <title>MoviCol</title>
 </head>
 
@@ -20,10 +20,9 @@
     </div>
   </header>
 
-  <button class="prev" type="button">&lt;</button>
-  <button class="next" type="button">&gt;</button>
-  
   <div class="cardList">
+    <button class="prev" type="button">&lt;</button>
+    <button class="next" type="button">&gt;</button>
     <div class="slider">
       
     </div>
