+++ css/index.css
@@ -1,3 +1,9 @@
+:root {
+  --color-main: #A1C4EB;
+  --color-back: #212121;
+  --color-card: #111111;
+}
+
 * {
   margin: 0;
   padding: 0;
@@ -6,7 +12,7 @@
 body {
   width: 1900px;
   margin: 0 auto;
-  background-color: #212121;
+  background-color: var(--color-back);
 }
 
 body::-webkit-scrollbar {
@@ -15,11 +21,11 @@ body::-webkit-scrollbar {
 }
 
 body::-webkit-scrollbar-track {
-  background-color: #212121;
+  background-color: var(--color-back);
 }
 
 body::-webkit-scrollbar-thumb {
-  background-color: #A1C4EB;
+  background-color: var(--color-main);
   border-radius: 1px;
 }
 
@@ -27,7 +33,7 @@ header {
   position: sticky;
   top: 0px;
   display: flex;
-  background-color: #212121;
+  background-color: var(--color-back);
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
@@ -37,7 +43,7 @@ header {
 }
 
 header>h1 {
-  color: #A1C4EB;
+  color: var(--color-main);
   padding: 10px 25px 0 25px;
 }
 
@@ -47,26 +53,26 @@ header>h1 {
   padding: 10px 25px 0 25px;
 }
 
-#index {
-  color: #A1C4EB;
+#input {
+  color: var(--color-main);
   font-size: 15px;
   padding-left: 7px;
-  background-color: #212121;
-  border: 2px solid #A1C4EB;
+  background-color: var(--color-back);
+  border: 2px solid var(--color-main);
   border-right: 0;
   width: 400px;
   height: 30px;
 }
 
-#index:focus {
+#input:focus {
   outline: none;
 }
 
 #inputButton {
-  background-color: #A1C4EB;
-  border: 2px solid #A1C4EB;
+  background-color: var(--color-main);
+  border: 2px solid var(--color-main);
   width: 50px;
-  color: #212121;
+  color: var(--color-back);
   font-size: 15px;
   font-weight: bolder;
 }
@@ -79,7 +85,7 @@ header>h1 {
 }
 
 .card {
-  background-color: #111111;
+  background-color: var(--color-card);
   border-radius: 15px;
   overflow: hidden;
   /* background-color: #212121; */
@@ -87,14 +93,14 @@ header>h1 {
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  border: 2px solid #212121;
+  border: 2px solid var(--color-back);
   width: 346px;
   height: 850px;
   margin-bottom: 25px;
 }
 
 .card:hover {
-  border: 2px solid #A1C4EB;
+  border: 2px solid var(--color-main);
   border-radius: 0;
   width: 346px;
   cursor: pointer;
@@ -106,7 +112,7 @@ header>h1 {
 }
 
 .card h2 {
-  color: #A1C4EB;
+  color: var(--color-main);
   margin-top: 5px;
   text-align: center;
 }
@@ -118,7 +124,7 @@ header>h1 {
 }
 
 .overview::-webkit-scrollbar {
-  background-color: #111111;
+  background-color: var(--color-card);
   /* background-color: #212121; */
 }
 
@@ -126,5 +132,5 @@ header>h1 {
   margin-top: 10px;
   margin-bottom: 5px;
   margin-left: 10px;
-  color: #A1C4EB;
+  color: var(--color-main);
 }
