+++ index.html
@@ -4,8 +4,8 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="./asset/css/index.css">
-  <script src="./asset/js/index.js"></script>
+  <link rel="stylesheet" href="./css/index.css">
+  <script type="module" src="./js/index.js"></script>
   <title>MoviCol</title>
 </head>
 
@@ -13,10 +13,10 @@
   <header>
     <h1>MoviCol</h1>
 
-    <div class="inputForm">
-      <input type="text" id="index" autofocus>
+    <form class="inputForm">
+      <input type="text" id="input">
       <button id="inputButton">검색</button>
-    </div>
+    </form>
   </header>
 
   <section class="cardList">
